+++ server/src/bot/bots/Greedy.ts
@@ -1,7 +1,8 @@
 // Hi, I'm greedy,
 // I care about nothing but about getting balls into the house
 
-import { AiData, AiInterface, getMovesFromCards } from '../simulation/simulation'
+import { AiData } from '../simulation/output'
+import { AiInterface, getMovesFromCards } from '../simulation/simulation'
 
 export class Greedy implements AiInterface {
   choose(data: AiData) {
