+++ server/src/bot/bots/Raindom.ts
@@ -2,7 +2,8 @@
 // I can do a lot of things, like:
 // - Just playing randomly
 
-import { AiData, AiInterface, getMovesFromCards } from '../simulation/simulation'
+import { AiData } from '../simulation/output'
+import { AiInterface, getMovesFromCards } from '../simulation/simulation'
 
 export class Raindom implements AiInterface {
   choose(data: AiData) {
